reduced app.jsx, implemented main pages for each user type, more healthy structure, routing is to be done
