+++ frontend/my-app/src/containers/Login.jsx
@@ -3,8 +3,7 @@ import { FormGroup, FormControl, FormLabel } from "react-bootstrap";
 import LoaderButton from "../components/LoaderButton";
 import { useFormFields } from "../libs/hooksLib";
 import "./Login.css";
-import { BrowserRouter as Router, Switch, Route, Link } from "react-router-dom";
-import App from "../App";
+import axios from "axios";
 
 export default function Login({ loginSuccess, api, ...props }) {
   const [isLoading, setIsLoading] = useState(false);
@@ -29,7 +28,16 @@ export default function Login({ loginSuccess, api, ...props }) {
       .then(response => {
         //console.log(response);
         if (response.statusText === "OK") {
-          loginSuccess(response.data.user_id, response.data.token);
+          //loginSuccess(response.data.user_id, response.data.token);
+          localStorage.setItem("userId",response.data.user_id)
+          localStorage.setItem("userToken",response.data.token)
+          axios
+          .get("http://8.209.81.242:8000/users/" + response.data.user_id, {
+            headers: { Authorization: `Token ${response.data.token}` }
+          })
+          .then(res => {
+            localStorage.setItem("userGroup", res.data.groups[0])
+          });
         }
       })
       .catch(function(error) {
